+++ NEWS.md
@@ -3,6 +3,8 @@
 * Load GTEx data (`loadGtexData()`):
     - Support for loading GTEx data from V8, V7, V6 or V4 releases
     - Organise GTEx data into folders named based on release version
+* `plotLibrarySize()` added to plot library size from gene expression data and
+fixed library size plot not working properly in visual interface
 
 ## Bug fixes and minor changes